first

+++ DSA-1/algorithmAnalysis/findClosest.h
@@ -3,7 +3,7 @@
 
 /// Возвращает элемент массива arr размера n, наиболее близкий к заданному числу x.
 /// Если в массиве несколько элементов с одинаковой минимальной разницей, возвращается первый найденный.
-int findClosestElement(const int* arr, size_t n, int x);
+float findClosestElement(const float* arr, size_t n, float x);
 
 /// Функция для тестирования функции findClosestElement.
 void testFindClosestElement();
